Add a new file for ClientManager
